+++ synapseclient/__main__.py
@@ -14,6 +14,7 @@ import json
 import getpass
 import csv
 import re
+import shutil
 
 import synapseclient
 import synapseutils
@@ -394,6 +395,87 @@ def storeTable(args, syn):
     syn.logger.info('{"tableId": "%s"}', table_ent.tableId)
 
 
+def _replace_existing_config(path, auth_section):
+    # insert the auth section into the existing config file
+
+    # always make a backup of the existing config file,
+    # but don't overwrite an existing backup
+    i = 1
+    cur_config_path = os.path.dirname(os.path.realpath(path))
+    while True:
+        copy_filename = f"{path}.backup{i if i > 1 else ''}"
+        copy_to = os.path.join(cur_config_path, copy_filename)
+        if not os.path.exists(copy_to):
+            shutil.copyfile(path, copy_to)
+            break
+
+        i += 1
+
+    # find the existing authentication section, we'll replace it wholly with an updated
+    # section with the new values.
+    with open(path, 'r') as config_o:
+        config_text = config_o.read()
+
+    matcher = re.search(
+        r'^[ \t]*(\[authentication\].*?)(^[ \t]*\[|\Z)',
+        config_text,
+        flags=re.MULTILINE | re.DOTALL
+    )
+    if matcher:
+        # we matched an existing authentication section
+        new_config_text = config_text[:matcher.start(1)] + auth_section + config_text[matcher.end(1):]
+
+    else:
+        # weren't able to find an authentication section so
+        # we write a new authentication section at the start of the file
+        new_config_text = auth_section + '\n\n' + config_text
+
+    return new_config_text
+
+
+def _generate_new_config(auth_section):
+    # insert the auth section into the default config template file
+
+    script_dir = os.path.dirname(os.path.realpath(__file__))
+
+    with open(os.path.join(script_dir, ".synapseConfig"), "r") as config_o:
+        config_text = config_o.read()
+
+    # Replace text in configuration
+    new_config_text = re.sub(
+        r'#\[authentication\].*<authtoken>',
+        auth_section,
+        config_text,
+        flags=re.MULTILINE | re.DOTALL
+    )
+    return new_config_text
+
+
+def config(args, syn):
+    """Create/modify a synapse configuration file"""
+    user, secret = _prompt_for_credentials()
+
+    # validate the credentials provided and determine what login
+    # mechanism was used (password, authToken, apiKey)
+    # this means that writing a config requires connectivity
+    # (and that the endpoints in the current config point to
+    # the endpoints of the credentials)
+    login_key = _authenticate_login(syn, user, secret, silent=True)
+
+    auth_section = '[authentication]\n'
+    if user:
+        auth_section += f"username={user}\n"
+    auth_section += f"{login_key.lower()}={secret}\n\n"
+
+    if os.path.exists(args.configPath):
+        config_text = _replace_existing_config(args.configPath, auth_section)
+    else:
+        config_text = _generate_new_config(auth_section)
+
+    with open(args.configPath, "w") as config_o:
+        config_o.write(config_text)
+
+
 def submit(args, syn):
     """
     Method to allow challenge participants to submit to an evaluation queue.
@@ -441,6 +523,12 @@ def submit(args, syn):
     submission = syn.submit(args.evaluationID, args.entity, name=args.name, team=args.teamName)
     sys.stdout.write('Submitted (id: %s) entity: %s\t%s to Evaluation: %s\n'
                      % (submission['id'], submission['entityId'], submission['name'], submission['evaluationId']))
+
+
+def get_download_list(args, syn):
+    """Download files from the Synapse download cart"""
+    manifest_path = syn.get_download_list(downloadLocation=args.downloadLocation)
+    syn.logger.info(f"Manifest file: {manifest_path}")
 
 
 def login(args, syn):
@@ -629,8 +717,14 @@ def build_parser():
                              help='Send notifications via Synapse messaging (email) at specific intervals, '
                                   'on errors and on completion.')
     parser_sync.add_argument('--retries', metavar='INT', type=int, default=4)
-    parser_sync.add_argument('manifestFile', metavar='FILE', type=argparse.FileType("r"),
-                             help='A tsv file with file locations and metadata to be pushed to Synapse.')
+    parser_sync.add_argument(
+        'manifestFile', metavar='FILE', type=argparse.FileType("r"),
+        help=(
+            'A tsv file with file locations and metadata to be pushed to Synapse. '
+            'See https://python-docs.synapse.org/build/html/synapseutils.html#synapseutils.sync.syncToSynapse '
+            'for details on the format of a manifest.'
+        )
+    )
     parser_sync.set_defaults(func=sync)
 
     parser_store = subparsers.add_parser('store',  # Python 3.2+ would support alias=['store']
@@ -752,6 +846,12 @@ def build_parser():
                            help='Do not copy the wiki pages')
     parser_cp.set_defaults(func=copy)
 
+    parser_get_dl_list = subparsers.add_parser('get-download-list',
+                                               help='Download files from the Synapse download cart')
+    parser_get_dl_list.add_argument('--downloadLocation', metavar='path', type=str, default="./",
+                                    help='Directory to download file to [default: %(default)s].')
+    parser_get_dl_list.set_defaults(func=get_download_list)
+
     parser_associate = subparsers.add_parser('associate',
                                              help=(
                                                  'Associate local files with the files stored in Synapse so that calls'
@@ -841,6 +941,10 @@ See https://docs.synapse.org/rest/org/sagebionetworks/repo/web/controller/TableE
     parser_list.add_argument('-m', '--modified', action='store_true', default=False, required=False,
                              help='List modified by and modified date')
     parser_list.set_defaults(func=ls)
+
+    parser_config = subparsers.add_parser('config',
+                                          help='Create or modify a Synapse configuration file')
+    parser_config.set_defaults(func=config)
 
     parser_set_provenance = subparsers.add_parser('set-provenance',
                                                   help='create provenance records')
@@ -946,7 +1050,7 @@ See https://docs.synapse.org/rest/org/sagebionetworks/repo/web/controller/TableE
     parser_login.add_argument('-u', '--username', dest='synapseUser',
                               help='Username used to connect to Synapse')
     parser_login.add_argument('-p', '--password', dest='synapsePassword',
-                              help='Password or api key used to connect to Synapse')
+                              help='This will be deprecated. Password or api key used to connect to Synapse.')
     parser_login.add_argument('--rememberMe', '--remember-me', dest='rememberMe', action='store_true', default=False,
                               help='Cache credentials for automatic authentication on future interactions with Synapse')
     parser_login.set_defaults(func=login)
@@ -1006,7 +1110,7 @@ See https://docs.synapse.org/rest/org/sagebionetworks/repo/web/controller/TableE
 
 def perform_main(args, syn):
     if 'func' in args:
-        if args.func != login:
+        if args.func != login and args.func != config:
             login_with_prompt(syn, args.synapseUser, args.synapsePassword, silent=True)
         try:
             args.func(args, syn)
@@ -1027,24 +1131,31 @@ def login_with_prompt(syn, user, password, rememberMe=False, silent=False, force
         _authenticate_login(syn, user, password, silent=silent, rememberMe=rememberMe, forced=forced)
     except SynapseNoCredentialsError:
         # there were no complete credentials in the cache nor provided
-        if not user:
-            # if username was passed then we use that username
-            user = input("Synapse username (leave blank if using an auth token): ")
+        user, passwd = _prompt_for_credentials(user=user)
 
-        # if no username was provided then prompt for auth token, since no other secret will suffice without a user
-        secret_prompt = f"Password, api key, or auth token for user {user}:" if user else "Auth token:"
-
-        passwd = None
-        while not passwd:
-            # if the terminal is not a tty, we are unable to read from standard input
-            # For git bash using python getpass
-            # https://stackoverflow.com/questions/49858821/python-getpass-doesnt-work-on-windows-git-bash-mingw64
-            if not sys.stdin.isatty():
-                raise SynapseAuthenticationError(
-                    "No password, key, or token was provided and unable to read from standard input")
-            else:
-                passwd = getpass.getpass(secret_prompt)
         _authenticate_login(syn, user, passwd, rememberMe=rememberMe, forced=forced)
+
+
+def _prompt_for_credentials(user=None):
+    if not user:
+        # if username was passed then we use that username
+        user = input("Synapse username (leave blank if using an auth token): ")
+
+    # if no username was provided then prompt for auth token, since no other secret will suffice without a user
+    secret_prompt = f"Password, api key, or auth token for user {user}:" if user else "Auth token:"
+
+    passwd = None
+    while not passwd:
+        # if the terminal is not a tty, we are unable to read from standard input
+        # For git bash using python getpass
+        # https://stackoverflow.com/questions/49858821/python-getpass-doesnt-work-on-windows-git-bash-mingw64
+        if not sys.stdin.isatty():
+            raise SynapseAuthenticationError(
+                "No password, key, or token was provided and unable to read from standard input")
+        else:
+            passwd = getpass.getpass(secret_prompt)
+
+    return user, passwd
 
 
 def _authenticate_login(syn, user, secret, **login_kwargs):
@@ -1078,7 +1189,7 @@ def _authenticate_login(syn, user, secret, **login_kwargs):
                 login_kwargs_with_secret = {login_key: secret} if login_key else {}
                 login_kwargs_with_secret.update(login_kwargs)
                 syn.login(user, **login_kwargs_with_secret)
-                break
+                return login_key
             except SynapseNoCredentialsError:
                 # SynapseNoCredentialsError is a SynapseAuthenticationError but we don't want to handle it here
                 raise