+++ tests/unit/synapseclient/unit_test_tables.py
@@ -20,11 +20,12 @@ import synapseclient.table
 from synapseclient.table import Column, Schema, CsvFileTable, TableQueryResult, cast_values, \
     as_table_columns, Table, build_table, RowSet, SelectColumn, EntityViewSchema, RowSetTable, Row, PartialRow, \
     PartialRowset, SchemaBase, _get_view_type_mask_for_deprecated_type, EntityViewType, _get_view_type_mask, \
-    MAX_NUM_TABLE_COLUMNS, SubmissionViewSchema, escape_column_name, join_column_names
+    MAX_NUM_TABLE_COLUMNS, SubmissionViewSchema, escape_column_name, join_column_names, MaterializedViewSchema, \
+    Dataset
 
 from synapseclient.core.utils import from_unix_epoch_time
 from unittest.mock import patch
-from collections import OrderedDict
+from collections import OrderedDict, abc
 
 
 def test_cast_values():
@@ -147,6 +148,71 @@ def test_schema():
     assert len(schema.columns_to_store) == 3
     assert Column(name='Living', columnType='BOOLEAN') not in schema.columns_to_store
     assert Column(name='Hipness', columnType='DOUBLE') in schema.columns_to_store
+
+
+def test_materialized_view():
+    """Test creation of materialized view
+    """
+    mat_view = MaterializedViewSchema(
+        name='My Table',
+        parent="syn1000001",
+        definingSQL="SELECT * FROM syn111 F JOIN syn2222 P on (F.patient_id = P.patient_id)"
+    )
+    mat_view._synapse_entity_type == "org.sagebionetworks.repo.model.table.MaterializedView"
+
+    assert not mat_view.has_columns()
+
+    mat_view.addColumn(Column(id='1', name='Name', columnType='STRING'))
+
+    assert mat_view.has_columns()
+    assert mat_view.properties.columnIds == ['1']
+
+    mat_view.removeColumn('1')
+    assert not mat_view.has_columns()
+    assert mat_view.properties.columnIds == []
+
+
+def test_dataset():
+    dataset = Dataset(
+        name="Pokedex",
+        parent="syn123",
+        dataset_items=[{'entityId': "syn111", 'versionNumber': 1}]
+    )
+    dataset._synapse_entity_type == "org.sagebionetworks.repo.model.table.Dataset"
+
+    # Items in the dataset should be captured in the `dataset_items` property,
+    # not `items`. e.items should continue to be an ItemsView().
+    assert hasattr(dataset, 'datasetItems')
+    assert isinstance(dataset.datasetItems, list)
+    assert isinstance(dataset.items(), abc.ItemsView)
+
+    # Default column IDs are used when schema is not specified.
+    assert not dataset.has_columns()
+    assert dataset.has_item("syn111") is True
+    assert dataset.has_item("syn222") is False
+    assert len(dataset) == 1
+
+    # added item must be a dictionary with two keys: entityId, versionNumber
+    with pytest.raises(ValueError):
+        dataset.add_item("syn222")
+    with pytest.raises(LookupError):
+        dataset.add_item({'entityId': '222'})
+
+    dataset.add_item({'entityId': 'syn222', 'versionNumber': 1})
+    assert dataset.has_item("syn222") is True
+    assert len(dataset) == 2
+
+    dataset.remove_item("syn222")
+    assert dataset.has_item("syn222") is False
+    assert len(dataset) == 1
+
+    with pytest.raises(ValueError):
+        dataset.add_item({'entityId': "syn111", 'versionNumber': 2}, force=False)
+    dataset.add_item({'entityId': "syn111", 'versionNumber': 2})
+    assert len(dataset) == 1
+
+    dataset.empty()
+    assert len(dataset) == 0
 
 
 def test_RowSetTable():
@@ -343,6 +409,17 @@ def test_pandas_to_table_convert__integer_data_type():
     test_table = Table(schema, test_df)
     for col in test_table.headers:
         assert col.columnType == "INTEGER"
+
+
+def test_pandas_to_table_convert__mixed_int_float_data_type():
+    test_df = pd.DataFrame({'int1': [1, 2, 990]})
+    schema = Schema(name="Foo", parent="syn12345")
+    # Since I can't seem to define the "mixed-integer-float" type, I force `infer_dtype`
+    # to return "mixed-integer-float" to test this feature.
+    with patch.object(pd.api.types, "infer_dtype", return_value="mixed-integer-float"):
+        test_table = Table(schema, test_df)
+        for col in test_table.headers:
+            assert col.columnType == "DOUBLE"
 
 
 def test_pandas_to_table_convert__boolean_data_type():